remove absolute positioning & fix displaying figures
It seems like kitty changed their handling of images somewhat.

The result is that the initial scrolling issues that lead to
prefering absolute positioning seem to be fixed, which means
we can simply use --align=left now.

Another side effect was that display of images was broken at
some point. This should be fixed as well with the above change.

Lastly, I've dropped transparency, as cell content now flows
properly around images, even in scrollback.