enable packaging via setuptools
I couldn't get setuptools to use the root directory as
the source of the package, so I've moved everything to a
sub-folder, which is a workflow with better documentation.

A symlink is left behind to avoid breaking the setup for
people that install this to their pythonpath via git clone.