update the interactive repl heuristic
The current heuristic doesn't detect ipython as an
interactive environment. In addition, forgetting
the `-i` during startup is a common UX error.

Because `sys.ps1` is documented to imply an interactive
environment, we can add it as an additional signal for our
heuristic. https://docs.python.org/3/library/sys.html#sys.ps1

A dataset gathered by the user `pyjamas` on StackOverflow
implies that this should work reasonably well:
https://stackoverflow.com/a/64523765